// hairlab/src/app/models/enums/reflection.ts
export enum Reflection {

  NATURAL = 'NATURAL',

  ASH = 'ASH',

  BLUE = 'BLUE',

  GREEN = 'GREEN',

  GOLD = 'GOLD',

  COPPER = 'COPPER',

  RED = 'RED',

  RED_VIOLET = 'RED_VIOLET',

  MAHOGANY = 'MAHOGANY',

  VIOLET = 'VIOLET',

  BROWN = 'BROWN',

  MOCHA = 'MOCHA',

  BEIGE = 'BEIGE',

  PEARL = 'PEARL',

  CENDRE = 'CENDRE',

  OTHER = 'OTHER'
}